requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
python-dotenv>=0.19.0
python-dateutil>=2.8.2
pytz>=2021.3

# Web Framework
streamlit>=1.10.0
flask>=2.0.0

# Data Visualization
matplotlib>=3.4.0
plotly>=5.3.0

# Machine Learning
tensorflow>=2.8.0
scikit-learn>=1.0.0
stable-baselines3>=1.5.0
gym>=0.21.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0

# Testing
pytest>=6.2.5

# Optional (for GPU support)
# tensorflow-gpu>=2.8.0
# cupy-cuda11x  # Select appropriate CUDA version
